requirements.txt
numpy
scikit-learn
joblib
flask
pillow